Add a currently disabled feature.
Make the sizeof pragma pack structures that same as microsoft compilers do it.
That is, don't add padding to the end of the structure.

Although it seems to work, and passed some testing, it needs regression testing and 
review by someone knowledgeable. Wouldn't be surprised if I got it all wrong.


git-svn-id: https://svn.code.sf.net/p/cegcc/code/trunk@37 d7810a3d-100a-0410-8641-c3624a9c11f1