The attached patch fills in the structures and constants needed for making Today Screen plug ins. Most constants and structures were obtained on the following links:
http://msdn2.microsoft.com/en-us/library/aa454291.aspx
http://blogs.msdn.com/mikehall/archive/2004/02/22/78141.aspx

The MAX_ITEMNAME constant was obtained by dumping the data pointer
passed by InitializeCustomItem() and looking at the offset of the
szDLLPath member (see the DUMP_DATA define in the code).
The dump.txt file contains the edited result of this dump.

The TODAYM_DRAWWATERMARK constant was guessed because it was the one
missing between the two other similar messages.

The WM_TODAYCUSTOM_CLEARCACHE and WM_TODAYCUSTOM_QUERYREFRESHCACHE
messages were found by dumping the unknown window messages received
greater that WM_USER.


git-svn-id: https://svn.code.sf.net/p/cegcc/code/trunk@871 d7810a3d-100a-0410-8641-c3624a9c11f1